Real valued rules evaluation implemented. Moved base rule evaluation to rules package
