+++ internal/common/client_test.go
@@ -1,44 +1,1 @@
 package common
-
-import (
-	"github.com/SOMAS2020/SOMAS2020/internal/common/rules"
-	"gonum.org/v1/gonum/mat"
-	"testing"
-)
-
-// TestBasicRuleEvaluatorPositive Checks whether rule we expect to evaluate as true actually evaluates as such
-func TestBasicRuleEvaluatorPositive(t *testing.T) {
-	result, err := BasicRuleEvaluator("Kinda Complicated Rule")
-	if !result {
-		t.Errorf("Rule evaluation came as false, when it was expected to be true, potential error with value '%v'", err)
-	}
-}
-
-// TestBasicRuleEvaluatorNegative Checks whether rule we expect to evaluate as false actually evaluates as such
-func TestBasicRuleEvaluatorNegative(t *testing.T) {
-	registerTestRule()
-	result, err := BasicRuleEvaluator("Kinda Test Rule")
-	if result || err != nil {
-		t.Errorf("Rule evaluation came as true, when it was expected to be false, potential error with value '%v'", err)
-	}
-}
-
-func registerTestRule() {
-
-	//A very contrived rule//
-	name := "Kinda Test Rule"
-	reqVar := []string{
-		"number_of_islands_contributing_to_common_pool",
-		"number_of_failed_forages",
-		"number_of_broken_agreements",
-		"max_severity_of_sanctions",
-	}
-
-	v := []float64{1, 0, 0, 0, -4, 0, -1, -1, 0, 2, 0, 0, 0, 1, -2, 0, 0, 1, 0, -1}
-	CoreMatrix := mat.NewDense(4, 5, v)
-	aux := []float64{0, 0, 0, 0}
-	AuxiliaryVector := mat.NewVecDense(4, aux)
-
-	rules.RegisterNewRule(name, reqVar, *CoreMatrix, *AuxiliaryVector)
-	// Check internal/clients/team3/client.go for an implementation of a basic evaluator for this rule
-}
